refactor: human 글자들 participant로 변경
