+++ README.md
@@ -72,7 +72,7 @@
 - [x] value에 값을 할당하는 로직도 init() 안에 넣어도 될 것 같네요.
 - [x] popCards 검증로직 개선
 - [x] INIT_NUMBER라는 이름만으로는 의미를 온전하게 전달하지 못하는 것 같아요.
-- [ ] human이라는 단어는 어디서 온 걸까요? participant와 혼용될 필요가 있나요?
+- [x] human이라는 단어는 어디서 온 걸까요? participant와 혼용될 필요가 있나요?
 
 2단계 피드백
