+++ src/test/java/blackjack/domain/players/PlayersTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.human;
+package blackjack.domain.players;
 
 import static blackjack.domain.Fixtures.ACE;
 import static blackjack.domain.Fixtures.TEN;
@@ -7,8 +7,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.Fixtures;
-import blackjack.domain.participant.Players;
-import blackjack.domain.participant.human.Dealer;
+import blackjack.domain.players.participant.Dealer;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
